+++ CHANGES.md
@@ -14,8 +14,20 @@ To be released.
 
 ### Added interfaces
 
+ -  Added `trustedStateValidators` option to `Swarm<T>.PreloadAsync()` method.
+    If any peer in this set is reachable and there is no built up blockchain
+    in a current node, `Swarm<T>` receives the latest states of the major
+    blockchain from that trusted peer, which is also calculated by that peer,
+    instead of autonomously calculating the states from scratch.
+    Note that this option is intended to be exposed to end users through
+    a feasible user interface so that they can decide whom to trust
+    for themselves.
+    [[#272], [#343]]
+
 ### Behavioral changes
 
+ -  `BlockChain<T>.PreloadAsync()` method became to omit rendering of
+    `IAction`s in the preloaded behind blocks.  [[#272], [#343]]
  -  `Swarm<T>` became to have two more message types: `GetRecentStates` (`0x0b`)
     and `RecentStates` (`0x0c`).  [[#272], [#343]]